# api/openapi.yaml
openapi: 3.1.0
info:
  title: HookRelay API
  version: 0.8.0-dev
  description: >-
    Producer requests are idempotent by key. Transient signed delivery
    failures are retried within a bounded attempt and age budget. Expired
    worker claims leave an unknown attempt before recovery. Public HTTPS
    targets require a configured, versioned key ID. The demo HTTP exception
    requires an explicit local demo profile. This is the local v0.8 contract;
    the API schema is the same in direct Go, Compose, and kind runs.
servers:
  - url: http://127.0.0.1:8080
    description: Direct local API, Compose host port, or kind loopback port-forward
components:
  securitySchemes:
    producerToken:
      type: http
      scheme: bearer
    adminToken:
      type: http
      scheme: bearer
  schemas:
    Endpoint:
      type: object
      required: [id, name, url, key_id, version, enabled, created_at]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        url: {type: string, format: uri}
        key_id: {type: string, description: Signing key identifier; never the key value}
        version: {type: integer, minimum: 1}
        enabled: {type: boolean}
        created_at: {type: string, format: date-time}
    EndpointInput:
      type: object
      additionalProperties: false
      required: [name, url]
      properties:
        name: {type: string, pattern: '^[A-Za-z][A-Za-z0-9_-]{0,63}$'}
        url:
          type: string
          format: uri
          description: Public HTTPS on port 443, or one exact local demo URL in demo profile
        key_id:
          type: string
          description: Required for HTTPS targets; must name a configured external key file
...
    EndpointPage:
      type: object
      required: [items, limit, offset]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Endpoint'}}
        limit: {type: integer, minimum: 1, maximum: 100}
        offset: {type: integer, minimum: 0, maximum: 10000}
        next_offset: {type: integer, minimum: 1}
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
paths:
...
  /v1/endpoints:
    post:
      summary: Create a managed target
      security: [{adminToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EndpointInput'}
      responses:
        '201':
          description: Target created
          headers:
            Location: {description: Relative target URL, schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Endpoint'}
        '400': {description: Invalid name or URL, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '401': {description: Admin token required, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '409': {description: Name already exists, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '429': {description: Admin mutation rate exceeded, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '413': {description: Body exceeds 2 KiB, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '415': {description: Content-Type must be application/json, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '503': {description: Target could not be stored, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    get:
      summary: List managed targets, including disabled targets
      security: [{adminToken: []}]
      parameters:
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, maximum: 10000, default: 0}
      responses:
        '200': {description: Target page, content: {application/json: {schema: {$ref: '#/components/schemas/EndpointPage'}}}}
        '400': {description: Invalid pagination, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '401': {description: Admin token required, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '503': {description: Targets could not be read, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...